+++ source/MaterialXTest/Element.cpp
@@ -1,0 +1,58 @@
+//
+// TM & (c) 2017 Lucasfilm Entertainment Company Ltd. and Lucasfilm Ltd.
+// All rights reserved.  See LICENSE.txt for license.
+//
+
+#include <MaterialXTest/Catch/catch.hpp>
+
+#include <MaterialXCore/Document.h>
+
+namespace mx = MaterialX;
+
+TEST_CASE("Element", "[element]")
+{
+    // Create a document.
+    mx::DocumentPtr doc = mx::createDocument();
+
+    // Create elements
+    mx::ElementPtr elem1 = doc->addChildOfCategory("generic");
+    mx::ElementPtr elem2 = doc->addChildOfCategory("generic");
+    REQUIRE(elem1->getParent() == doc);
+    REQUIRE(elem2->getParent() == doc);
+    REQUIRE(elem1->getRoot() == doc);
+    REQUIRE(elem2->getRoot() == doc);
+    REQUIRE(doc->getChildren()[0] == elem1);
+    REQUIRE(doc->getChildren()[1] == elem2);
+
+    // Set hierarchical properties
+    doc->setFilePrefix("folder/");
+    doc->setColorSpace("lin_rec709");
+    REQUIRE(elem1->getActiveFilePrefix() == doc->getFilePrefix());
+    REQUIRE(elem2->getActiveColorSpace() == doc->getColorSpace());
+
+    // Modify element names.
+    elem1->setName("elem1");
+    elem2->setName("elem2");
+    REQUIRE(elem1->getName() == "elem1");
+    REQUIRE(elem2->getName() == "elem2");
+    REQUIRE_THROWS_AS(elem2->setName("elem1"), mx::Exception);
+
+    // Modify element order.
+    mx::DocumentPtr doc2 = doc->copy();
+    REQUIRE(*doc2 == *doc);
+    doc2->setChildIndex("elem1", doc2->getChildIndex("elem2"));
+    REQUIRE(*doc2 != *doc);
+    doc2->setChildIndex("elem1", doc2->getChildIndex("elem2"));
+    REQUIRE(*doc2 == *doc);
+    REQUIRE_THROWS_AS(doc2->setChildIndex("elem1", 100), mx::Exception);
+    REQUIRE(*doc2 == *doc);
+
+    // Create and test an orphaned element.
+    mx::ElementPtr orphan;
+    {
+        mx::DocumentPtr doc3 = doc->copy();
+        orphan = doc3->getChild("elem1");
+        REQUIRE(orphan);
+    }
+    REQUIRE_THROWS_AS(orphan->getDocument(), mx::ExceptionOrphanedElement);    
+}